Attempted to fix a bug with collapsed bridge detection
The bots automatically detect if they are on of the bridges that fall apart based on the amount of weght on it. There was an issue with the old system used where bots would believe they were on a broken bridge and get stuck forever. This only really happend on Adamance and rarely Vow.